refactor(temperature): move app-ready startup into service
- fold the post-ready temperature initialization handler into TemperatureService

- before this change the temperature module registered a separate listener provider only to forward app.ready into initializeProviders

- that added module-level indirection without adding real behavior or isolation

- the service now owns its own app-ready handler and logs startup initialization failures directly

- the standalone startup-task files were removed and the service spec now covers the post-ready path
